split commands.go into multiple files
