+++ Makefile
@@ -12,13 +12,18 @@ BIN = $(BUILD_DIR)/hermit-$(GOOS)-$(GOARCH)
 
 all: lint test build
 
-lint:
+lint: ## run golangci-lint
 	./bin/golangci-lint run
 
-test:
+test: ## run tests
 	./bin/go test -v ./...
 
-build:
+build: ## builds binary and gzips it
 	mkdir -p build
 	CGO_ENABLED=0 go build -ldflags "-X main.version=$(VERSION) -X main.channel=$(CHANNEL)" -o $(BIN) ./cmd/hermit
 	gzip -c9 $(BIN) > $(BIN).gz
+
+help: ## Display this help message
+	@awk 'BEGIN {FS = ":.*?## "} /^[a-zA-Z_\/-]+:.*?## / {printf "\033[34m%-30s\033[0m %s\n", $$1, $$2}' $(MAKEFILE_LIST) | \
+  	sort | \
+  	grep -v '#'
